Agrego getSocialNetwork a utils
